Add ForwardLine
Implementing forward line uncovered an error in Buffer.FindForwards.
Unable to fix the error quickly and in anticipation of future
functionality, such as regexp based searches, caching of the buffer's
contents as a string is used instead:

1. the initial contents of the buffer are stored in a cache on the
   editor.  The cache is marked as clean.
2. on any operation that changes the contents of the underlying
   buffer, the cache is marked as dirty.
3. when requesting the buffer's contents by calling `Editor.String()`,
   the cached contents are returned.  If the cache is marked as dirty,
   the cache is updated by calling `buffer.String()`.

This choice has mainly been made to keep the code simple and avoids
having to reimplement a large set of functions on the new buffer
type (regular expression matching, finding substrings, etc).